+++ res/resource.h
@@ -254,6 +254,7 @@
 #define IDM_FILE_FILL_FORM       7420
 #define IDM_FILE_SIGN_PDF        7421
 #define IDM_FILE_SIGN_CERT       7422
+#define IDM_FILE_TYPE_ON_PDF     7423
 #define IDD_COMMENTS             1160
 #define IDC_COMMENTS_LIST        1161
 #define IDC_COMMENTS_DELETE      1162